+++ src/lib/i18n/messages/en.json
@@ -8,6 +8,9 @@
 	"contribute_link": "Contribute",
 	"producers_link": "Producers",
 	"prices_link": "Prices",
+	"login_link": "Login",
+	"logout_link": "Logout",
+	"account_link": "Account",
 	"world_option": "World",
 	"github_link": "GitHub",
 	"search": {
