+++ src/lib/stores/pkceLoginStore.ts
@@ -1,5 +1,6 @@
 export const keycloak = 'http://auth.openfoodfacts.localhost:5600/realms/open-products-facts';
 export const pkceClientId = 'OFF_EXP_DEV';
+export const accountUrl = `${keycloak}/account/#/`;
 
 export type AuthTokens = {
 	expires_in: number;
@@ -16,7 +17,7 @@ export async function getAccessToken(useRefreshToken: boolean = false) {
 	const body = new URLSearchParams({
 		code: code ?? '',
 		grant_type: 'authorization_code',
-		redirect_uri: 'http://localhost:5173/callback', // changed to match login redirect_uri
+		redirect_uri: 'http://localhost:5173/login_callback', // changed to match login redirect_uri
 		client_id: pkceClientId,
 		code_verifier: verifier ?? ''
 	});
