+++ src/routes/login/+page.svelte
@@ -31,7 +31,7 @@
 		const lang = 'en';
 
 		// const redirectUri = `http://localhost:5604/${page}.html?clientId=${clientId}&pkceClientId=${pkceClientId}&clientSecret=${clientSecret}&lang=${lang}&keycloak=${encodeURIComponent(keycloak)}&mode=pkce`;
-		const redirectUri = 'http://localhost:5173/callback';
+		const redirectUri = 'http://localhost:5173/login_callback';
 
 		const url = `${keycloak}/protocol/openid-connect/auth?response_type=code&client_id=${pkceClientId}&redirect_uri=${encodeURIComponent(redirectUri)}&scope=openid+profile+offline_access&state=${nonce}&ui_locales=${lang}&code_challenge=${codeChallenge}&code_challenge_method=S256`;
 		window.location.href = url;
